Add useDocumentTitle hook for dynamically updating document title
This commit adds a new hook called useDocumentTitle that allows for dynamically updating the document title based on the provided title prop. If no title is provided, the default title (defined in the constants file) is used. The hook uses the useEffect hook to update the document title and also cleans up the effect on unmount.