Add spinner icon to login button
Added a spinner icon to the login button to indicate loading state. The
icon is displayed instead of the "Login" text when the button is
disabled due to missing email or password or when the login request is
being processed. The spinner icon is implemented as a new component in
SVGIcon.tsx.