Add remember user functionality to Login and Register pages
This commit adds the ability to remember a user on the Login and Register pages. The `remeberUser` function has been added to the `helpers.ts` file, which sets a "remember" key in local storage. This function is called when the user checks the "remember me" checkbox on the Login and Register pages. The `forgetUser` function has also been added to `helpers.ts`, which removes the "remember" key from local storage.