Refactor API request functions
This commit refactors the API request functions to accept an options object instead of separate arguments for URL, method, headers, data, and credentials. The options object is typed with the new `ApiRequestOptions` type. The `apiRequest` function now stringifies the data if it's an object and sets the content type header to JSON by default. The `privateRequest` function now sends the request with the access token in the headers instead of the options object. All API request functions have been updated to use the new options object format.